move filesystem-specific operations to filesystem module
